`Poller` now may interleave between IO and rendering
The `pollin` event in `vt100` may now interrupt its execution to give
a chance for the `screen` to be repainted, and then resume consuming
the buffer.

Commands with either very long or infinite output like `$ seq 100000;`,
`$ yes;`, and `$ tail -F my-log-file.txt;` are correctly displayed now.

There is still room for better settings & tuning, nonetheless this is a
step forward for a better poller/scheduler, which is still designed for
a single thread, but additional parallel resources may be added in the
future.

UTF-8 isn't properly handled in `vt100` as of now.