Performance notes
In terms of performance, the current code is at least one order of
magnitude slower of wayst.

A preliminary performance report indicates too much time is currently
spent into what it looks like opengl operations. On a stack flamegraph
analysis there is a big chunk of time spent on various `undefined`
symbols under a top-level `main:gdrv0` entry.

A proposal here is for default character rendering to write straight
into the framebuffer (as it used to be) and use renderbuffers for
scroll purposes only.

Additionally, after the cursor goes beyond the last line, it currently
repaints the screen "fully". That may be optimized if damages would
return the differences between two screens.